home page clean up WIP
