+++ src/app/components/AppWrapper/index.js
@@ -1,0 +1,3 @@
+import AppWrapper from './component';
+
+export default AppWrapper;
