+++ src/app/pages/index.js
@@ -1,0 +1,5 @@
+import HomePage from './Home';
+import LoginPage from './Login';
+import RegisterPage from './Register';
+
+export { HomePage, LoginPage, RegisterPage };
